pyproject.toml
[tool.poetry]
name = "ILLIXR-analysis"
version = "0.1.0"
description = ""
# Feel free to add yourself here.
# Perhaps we should populate this from the GitHub contributors.
authors = ["Samuel Grayson <sam@samgrayson.me>"]
license = "NCSA"
repository = "https://github.com/ILLIXR/ILLIXR-analysis"
packages = [
    { include = "illixr" },
]

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.3.2"}
typer-cli = "^0.0.11"
pandas = "^1.2.1"
attrs = "^20.3.0"
networkx = "^2.5"
anytree = "^2.8.0"
tqdm = "^4.56.0"
pygraphviz = "^1.7"
"charmonium.cache" = "^0.6.0"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^20.8b1"
mypy = "^0.800"
pylint = "^2.6.0"
ipython = "^7.19.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
